+++ semcore/data-table/__tests__/index.test.tsx
@@ -1,10 +1,11 @@
 import type { Intergalactic } from '@semcore/core';
 import { runDependencyCheckTests } from '@semcore/testing-utils/shared-tests';
 import { render, cleanup } from '@semcore/testing-utils/testing-library';
-import { expect, test, describe, beforeEach, vi, assertType } from '@semcore/testing-utils/vitest';
-import React, { useRef } from 'react';
+import { test, describe, beforeEach, vi, assertType } from '@semcore/testing-utils/vitest';
+import React from 'react';
 
 import { DataTable, UNIQ_ROW_KEY } from '../src';
+import type { CellRenderProps } from '../src/components/Body/Body.types';
 
 describe('data-table Dependency imports', () => {
   runDependencyCheckTests('data-table');
@@ -53,19 +54,34 @@ describe('DataTable', () => {
         />,
       );
     });
-    test('selectedRows typing', () => {
+    test('selectedRows/onSelectedRowsChange typing', () => {
+      const onSelectedRowsChange = (rows: string) => {};
+
       assertType<JSX.Element>(
         <DataTable
-          data={[{ id: 1, [UNIQ_ROW_KEY]: 'key' }]}
+          data={[{ id: 1 }]}
           aria-label='label'
           columns={[]}
-          selectedRows={['key']}
+          selectedRows={[1]}
+          uniqueRowKey='id'
           onSelectedRowsChange={(rows, e, opts) => {
             // eslint-disable-next-line @typescript-eslint/no-unused-expressions
-            rows; // string[]
+            rows; // number[]
             // eslint-disable-next-line @typescript-eslint/no-unused-expressions
             opts?.row.id; // should be number
           }}
+        />,
+      );
+      assertType<JSX.Element>(
+        <DataTable
+          data={[{ id: 1 }]}
+          aria-label='label'
+          columns={[]}
+          // @ts-expect-error
+          selectedRows={['1']}
+          uniqueRowKey='id'
+          // @ts-expect-error
+          onSelectedRowsChange={onSelectedRowsChange}
         />,
       );
     });
@@ -89,7 +105,7 @@ describe('DataTable', () => {
       );
 
       assertType<JSX.Element>(
-        <DataTable<{ id: number }[]>
+        <DataTable<{ id: number }[], 'id', number>
           data={[{ id: 1 }]}
           aria-label='label'
           columns={[]}
@@ -107,7 +123,7 @@ describe('DataTable', () => {
 });
 
 describe('DataTable.Cell', () => {
-  test('Should support ref via renderCell', () => {
+  test('Should support ref via renderCell', ({ expect }) => {
     const spy = vi.fn();
 
     const Test = () => {
@@ -135,5 +151,35 @@ describe('DataTable.Cell', () => {
 
     render(<Test />);
     expect(spy).toBeCalled();
+  });
+
+  // we have some error with rendering this example
+  test.skip('Should support rawData in custom renderCell function', ({ expect }) => {
+    const checkRowData = vi.fn();
+    const dataItem = { keyword: 'test', kd: '1', vol: null };
+
+    const RawDataTest = () => {
+      const customCellRender = (props: CellRenderProps<string | null>) => {
+        checkRowData(props.rawData);
+
+        return props.defaultRender();
+      };
+
+      return (
+        <DataTable
+          data={[dataItem]}
+          aria-label='table'
+          columns={[
+            { name: 'keyword', children: 'Keyword' },
+            { name: 'kd', children: 'KD' },
+            { name: 'vol', children: 'Vol.' },
+          ]}
+          renderCell={customCellRender}
+        />
+      );
+    };
+
+    render(<RawDataTest />);
+    expect(checkRowData).toBeCalledWith({ ...dataItem });
   });
 });