+++ semcore/base-components/CHANGELOG.md
@@ -2,11 +2,17 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [16.2.0] - 2025-07-24
+## [16.3.0] - 2025-08-09
 
 ### Changed
 
 - Styles for `:focus-visible` in Box and ScrollArea.
+
+## [16.2.0] - 2025-08-08
+
+### Added
+
+- `text-align` CSS property to `Box` component.
 
 ## [16.1.2] - 2025-07-23