+++ semcore/d3-chart/CHANGELOG.md
@@ -2,11 +2,17 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [16.2.0] - 2025-07-24
+## [16.2.0] - 2025-08-09
 
 ### Changed
 
 - Focus styles for `Plot`.
+
+## [16.1.1] - 2025-08-08
+
+### Fixed
+
+- Legend hides for single item even when `showLegend=true`.
 
 ## [16.1.0] - 2025-07-23