+++ semcore/icon/CHANGELOG.md
@@ -2,6 +2,23 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [16.6.0] - 2025-07-31
+
+### Added
+
+- New `CollapseList` icon.
+
+### Fixed
+
+- Removed fill from `TextItalic` size M icon.
+- Minified and removed fill from `BracketsCode` size L icon.
+
+## [16.5.0] - 2025-07-24
+
+### Added
+
+- New `AISEOToolkit` icon.
+
 ## [16.4.0] - 2025-07-21
 
 ### Added